Improve Pi Non-GPU-accel exposure (#148)
Changes exposure setters to accept a floating point input (rather than integer)
Updates the UI to change exposure in increments of 0.1 (rather than 1.0)
Updates quirky PI camera logic to use the raw_ interface with scaling/offset logic matching the GPU-accelerated pi3 camera logic from Declan.
Adds logic to disable auto-white-balance in the PI camera, which should yield more consistent vision processing results.

+++ photon-server/src/main/java/org/photonvision/vision/processes/VisionSourceSettables.java
@@ -42,7 +42,7 @@ public abstract class VisionSourceSettables {
         return configuration;
     }
 
-    public abstract void setExposure(int exposure);
+    public abstract void setExposure(double exposure);
 
     public abstract void setBrightness(int brightness);
