+++ photon-server/src/test/java/org/photonvision/vision/processes/VisionModuleManagerTest.java
@@ -69,7 +69,7 @@ public class VisionModuleManagerTest {
         }
 
         @Override
-        public void setExposure(int exposure) {}
+        public void setExposure(double exposure) {}
 
         @Override
         public void setBrightness(int brightness) {}
